src/auth/handlers.py
from datetime import datetime, timedelta
from typing import Optional

import jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from passlib.context import CryptContext
import os
from dotenv import load_dotenv
from bson import ObjectId

from src.models.schemas import User, UserInDB, TokenData, UserRole, users_collection

# Load environment variables
load_dotenv()

# JWT settings
SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key-for-jwt")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24  # 1 day

# Password hashing
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth2 scheme
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a password against a hash"""
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    """Hash a password"""
    return pwd_context.hash(password)


def get_user(email: str) -> Optional[UserInDB]:
    """Get a user by email"""
    user_dict = users_collection.find_one({"email": email})
    if user_dict:
        return UserInDB(**user_dict)
    return None


def authenticate_user(email: str, password: str) -> Optional[UserInDB]:
    """Authenticate a user"""
    user = get_user(email)
    if not user:
        return None
    if not verify_password(password, user.hashed_password):
        return None
    return user


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """Create a JWT access token"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


async def get_current_user(token: str = Depends(oauth2_scheme)) -> User:
    """Get the current user from a JWT token"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id: str = payload.get("sub")
        print(f"User ID: {user_id}")
        print(f"Payload: {payload}")
        if user_id is None:
            raise credentials_exception
        
        token_data = TokenData(user_id=user_id, exp=datetime.fromtimestamp(payload.get("exp")))
    except jwt.PyJWTError:
        raise credentials_exception
    
    user_dict = users_collection.find_one({"_id": token_data.user_id})
    if user_dict is None:
        raise credentials_exception
    
    # Convert ObjectId to string for the User model
    user_dict["_id"] = str(user_dict["_id"])
    
    return User(**user_dict)


async def get_current_active_user(current_user: User = Depends(get_current_user)) -> User:
    """Get the current active user"""
    if not current_user.is_active:
        raise HTTPException(status_code=400, detail="Inactive user")
    return current_user


async def get_current_admin_user(current_user: User = Depends(get_current_active_user)) -> User:
    """Get the current admin user"""
    if current_user.role != UserRole.ADMIN:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Admin access required"
        )
    return current_user


async def get_current_admin_user_with_edit_mode(current_user: User = Depends(get_current_admin_user)) -> User:
    """Get the current admin user with edit mode enabled"""
    if not current_user.settings.edit_mode_enabled:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Edit mode must be enabled for this operation"
        )
    return current_user


def check_edit_permission(user: User) -> bool:
    """Check if user has edit permissions"""
    return (
        user.role == UserRole.ADMIN and 
        user.settings.edit_mode_enabled and 
        user.is_active
    )


def check_admin_permission(user: User) -> bool:
    """Check if user has admin permissions"""
    return user.role == UserRole.ADMIN and user.is_active